net: add rvcnet paravirtualized network driver
rvc supports a paravirtualized network driver using an MMIO and CSR
based interface. This is a simple driver making use of it, based mostly
on loopback.c.

Probing happens via the device tree, the MMIO addresses for the RX and
TX buffers are read via custom CSRs. The IRQ is set statically to be
RV_IRQ_EXT (= SEIP = SupervisorExternalInterrupt). This would normally
only be the entrance point to indicate that some IRQ controller has
fired, but in our case there is no external CLINT, so the emulator uses
this pin directly for both rvcnet and UART. Currently, no distinction
between the two types is made, as the UART interrupt should not fire,
meaning every IRQ on this line will be treated as an rvcnet RX event.

The driver does not use TX queues, as the data is guaranteed to be sent
immediately upon writing to CSR_NET_TX_BUF_SIZE_AND_SEND. Similarly, RX
queues are not needed, as those are punted to the emulator
implementation, which will hold events until CSR_NET_RX_BUF_READY is
set again.